[WIP] cca_in_ad -> create cca templates and filter them
With this change, when logs_config.cca_in_ad is set, AD config providers
for containers will add a `container_collect_all` configuration template
for each container.

It then falls to the config reconciliation process to avoid
double-logging containers that have more-specific configuration.  This
is accomplished with a per-service template filter which drops the
container_collect_all configuration iff there is another template with a
nonempty LogsConfig.

TODO:
 - confirm that the form of the CCA config is correct